improve readability and update copyright notice
Improved readability by:
- Splitting the `linter.py` and formatters in separate files.
- Renamed `matches` to `problems`.
- Updated some variable names.

Also updated the copyright notices to 2020.
