+++ LICENSE
@@ -2,7 +2,7 @@ MIT License
 
 Copyright (c) 2013-2018 Will Thames <will@thames.id.au>
 Copyright (c) 2018 Ansible by Red Hat
-Modified work Copyright (c) 2019 Warpnet B.V.
+Modified work Copyright (c) 2020 Warpnet B.V.
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
