+++ saltlint/linter/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+# Copyright (c) 2020 Warpnet B.V.
+
+from saltlint.linter.collection import RulesCollection
+from saltlint.linter.match import Match
+from saltlint.linter.rule import Rule
+from saltlint.linter.runner import Runner
