+++ saltlint/linter/match.py
@@ -1,0 +1,18 @@
+# -*- coding: utf-8 -*-
+# Copyright (c) 2013-2014 Will Thames <will@thames.id.au>
+# Modified work Copyright (c) 2020 Warpnet B.V.
+
+
+class Match(object):
+
+    def __init__(self, linenumber, line, filename, rule, message=None):
+        self.linenumber = linenumber
+        self.line = line
+        self.filename = filename
+        self.rule = rule
+        self.message = message or rule.shortdesc
+
+    def __repr__(self):
+        formatstr = u"[{0}] ({1}) matched {2}:{3} {4}"
+        return formatstr.format(self.rule.id, self.message,
+                                self.filename, self.linenumber, self.line)
