+++ saltlint/rules/JinjaPillarGrainsGetFormatRule.py
@@ -1,13 +1,13 @@
 # -*- coding: utf-8 -*-
 # Copyright (c) 2016 Will Thames and contributors
 # Copyright (c) 2018 Ansible Project
-# Modified work Copyright (c) 2019 Warpnet B.V.
+# Modified work Copyright (c) 2020 Warpnet B.V.
 
 import re
-from saltlint.linter import SaltLintRule
+from saltlint.linter import Rule
 
 
-class JinjaPillarGrainsGetFormatRule(SaltLintRule):
+class JinjaPillarGrainsGetFormatRule(Rule):
     id = '211'
     shortdesc = 'pillar.get or grains.get should be formatted differently'
     description = "pillar.get and grains.get should always be formatted " \
