+++ saltlint/rules/LineTooLongRule.py
@@ -1,12 +1,12 @@
 # -*- coding: utf-8 -*-
 # Copyright (c) 2016 Will Thames and contributors
 # Copyright (c) 2018 Ansible Project
-# Modified work Copyright (c) 2019 Warpnet B.V.
+# Modified work Copyright (c) 2020 Warpnet B.V.
 
-from saltlint.linter import SaltLintRule
+from saltlint.linter import Rule
 
 
-class LineTooLongRule(SaltLintRule):
+class LineTooLongRule(Rule):
     id = '204'
     shortdesc = 'Lines should be no longer that 160 chars'
     description = (
