+++ tests/unit/TestNoIrregularSpacesRule.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+# Copyright (c) 2020 Warpnet B.V.
 
 import unittest
